Refactor the transit rolling cache.
Introduce a record type for the cache. Make it opaque to the rest of the world.

Then refactor the internals of the caching logic. In Erlang we can move a lot of decisions into pattern matches which simplifies the code flow quite a bit. Once this is done, we can excise some common functionality into helper functions.

Finally, introduce a `empty/0` call to create a new empty cache. This makes the module truly encapsulated. The notion of what a cache was was bleeding out of the module.