Use `exit/1` rather than `erlang:throw/1`.
The basic rule is that throws are expected to be caught by the surrounding code and it is never allowed to escape out of a subsystem/application. By changing throws into exits, we abort the process on failure, but we also obtain proper stack traces in the common test suites.