Remove matching on structures [{_, _} | _]
The original idea was to use this structure to handle proplist-encoded maps. But unfortunately, this leads to problems. An array containing a timestamp object has the form [{atom(), TS} | _] which matches the above. Deep encoding of composite maps fails at this since we end up encoding timestamp as atoms and lots of errors occur.

By disallowing the format, we reclaim full support for the current QuickCheck model.

Closes #17